POM review and improvement, and project cleaning.
POM was reviewed to update the dependencies of the project to the last
version and to improve the test execution.

Now it is possible to run the project with the testing database in a
Tomcat Server v7.0 using the mvn tomcat7:run goal. In addition, Selenium
tests now are executed as integration tests, starting the web server
before the test execution and stoping it after the tests end. Finally,
support for generating regular and cobertura test reports in HTML format
was also added.

The project was also cleaned by removing the MANIFEST.MF files and by
reallocating the SQL files.