Refactorizes the whole project for a simpler testing and execution
A big refactorization has been done in this commit, focused on
simplifying the testing and execution process. The Spring-DbUnit library
has been added to allow the use of DbUnit through annotations in the
test classes. This change has been complemented with two new test
listeners that allow the creation of custom initial contexts and the use
of custom SQL scripts to create and destroy the database tables before
and after the test execution, respectively. In addition, all the tests
are executed now using a HSQLDB database.

The datasource configuration has been pulled out of the project, and now
it must be provided by the container.

The sample classes have been reviewed, refactorized and documented with
JavaDoc.

The POM file has been reworked to ease test and application execution
using maven. The new execution modes are:
	- The default execution includes unit, integration and acceptance (with
Selenium) test execution.
	- No acceptance tests: mvn -P -acceptance-tests-cargo <phase>
	- Run server for manual acceptance tests: mvn
-Dcargo.tomcat.start.skip=true -Dcargo.tomcat.run.skip=false
-DskipTests=true pre-integration-test 
	- Run server with MySQL database: mvn -P
run-tomcat-mysql,-acceptance-tests-cargo -DskipTests=true package
cargo:run