Adds support for retrieving the data of an user
This commit adds the new resource UsersResource (/users) with a simple
functionality: retrieving the data of an user. This functionality
requires knowing the login of the user that is performing the request.
In this regard, this commit is an example of how this information can be
retrieved and how to test a resource that requires knowing the login of
an user.