Adds automatic deployment capabilities to Tomcat execution
The run-tomcat-mysql allows the execution of the project in a Tomcat 8 server
with a MySQL database through the "cargo:run" or "cargo:start" goals. This
commit adds the fizzed-watcher-maven-plugin to provide the project with
automatic redeployment when using the aforementioned profile.

For example, executing the Maven command the application will be served in a
Tomcat server an automatically redeployed when any file changes:

mvn -DskipTest=true -Prun-tomcat-mysql cargo:start fizzed-watcher:run