Fixes login check when user does no exists
When an invalid login is provided to the UsersDAO.get method it throws
an IllegalArgumentException. This may happen when the user's credentials
are checked by the LoginFilter, and the exception was not currently
managed. As a result, an user trying to login with an invalid login will
access the main.html page showing an error message.

This commit fixes this error, and invalid user login attempts now return
to the index.html page.