Fixes DB authentication error for MySQL 8.0.4+
The default authentication plugin for MySQL has been changed from
mysql_native_password to caching_sha2_password in version 8.0.4. This
causes that an access failure for the user created in the database
scripts when trying to login from Tomcat.

This commit changes the MySQL database scripts to maintain the
compatibility between Tomcat and MySQL in versions 8.0.4+.